% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_means.R
\docType{data}
\name{StatMeans}
\alias{StatMeans}
\title{The Stat object}
\description{
This \code{ggproto} class creates the geom_means and stat_means layer.
It is not intended to be used by user.
}
\keyword{datasets}
